Fix compile on unix
