+++ Cargo.toml
@@ -43,3 +43,6 @@ tracing = "0.1"
 tracing-subscriber = "0.3"
 webpki-roots = "0.22"
 regex = "1"
+
+[target.'cfg(unix)'.dependencies]
+hyperlocal = "0.8"
