+++ src/web/mod.rs
@@ -16,6 +16,13 @@ use std::net::SocketAddr;
 use thiserror::Error;
 use tokio_util::sync::CancellationToken;
 use tower_http::cors::{self, CorsLayer};
+#[cfg(unix)]
+use {
+    std::path::Path,
+    std::fs::Permissions,
+    std::os::unix::fs::PermissionsExt,
+    hyperlocal::UnixServerExt
+};
 
 #[derive(Clone, Copy)]
 pub struct WebAppData {
